Move form field to the separate folder
