+++ codex-ui/src/vue/form-section/index.ts
@@ -1,5 +1,4 @@
 export * from './FormSection.types.js';
 import FormSection from './FormSection.vue';
-import FormField from './FormField.vue';
 
-export { FormSection, FormField };
+export { FormSection };
