+++ codex-ui/src/vue/index.ts
@@ -1,5 +1,6 @@
 export * from './button';
-export * from './form';
+export * from './form-section';
+export * from './form-field';
 export * from './heading';
 export * from './input';
 export * from './row';
